requirements.txt
requests==2.26.0
Django==3.2
django_filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
pandas==2.2.3
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
isort==6.0.1
flake8==5.0.4